--- src/Entity/Trait/IdentiableTrait.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace App\Entity\Trait;
-
-use Doctrine\ORM\Mapping as ORM;
-
-trait IdentiableTrait
-{
-    #[ORM\Column(name: 'id', type: 'bigint', unique: true)]
-    #[ORM\Id]
-    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
-    private ?int $id = null;
-
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    public function setId(int $id): void
-    {
-        $this->id = $id;
-    }
-}
